feat: Modularização no crud de usuários
